Summary 13.01
First day playing around with VSC. Very challenging, yet very educational.

Started by setting up my HTML file and creating a CSS, and linked them together.
Made a hamburger menu for the header, and made it work for smaller and lager screen sizes.
Added the background image and text to the main part of my home page.
Created the HTML files for the about-page, contact-page and products-page, and linked them to the home-page.